+++ main/views/suite.py
@@ -1,5 +1,5 @@
 import json
-import traceback
+import logging
 from django.http import HttpResponseRedirect, JsonResponse, HttpResponse, HttpResponseBadRequest, Http404
 from django.shortcuts import render
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
@@ -12,6 +12,8 @@ from main.models import Suite, Case, SuiteVsCase
 from main.forms import PaginatorForm, SuiteForm
 from main.views.general import get_query_condition
 from django.forms.models import model_to_dict
+
+logger = logging.getLogger('django.request')
 
 
 # 用例列表
@@ -35,7 +37,11 @@ def suites(request):
         if keyword.strip() != '':
             keyword_list.append(keyword)
     q = get_query_condition(keyword_list)
-    objects = Suite.objects.filter(q, is_active=True).order_by('id').values('pk', 'name', 'keyword', 'config__name').annotate(m2m_count=Count('case'))
+    objects = Suite.objects.filter(q, is_active=True).order_by('id').values('pk', 'name', 'keyword', 'config__name')
+    objects2 = Suite.objects.filter(is_active=True, case__is_active=True).values('pk').annotate(m2m_count=Count('case'))
+    count_ = {o['pk']: o['m2m_count'] for o in objects2}
+    for o in objects:
+        o['m2m_count'] = count_.get(o['pk'], 0)
     paginator = Paginator(objects, size)
     try:
         objects = paginator.page(page)
@@ -55,7 +61,7 @@ def suite(request, pk):
     try:
         obj = Suite.objects.select_related('creator', 'modifier').get(pk=pk)
     except Suite.DoesNotExist:
-        raise Http404('Step does not exist')
+        raise Http404('Suite does not exist')
     if request.method == 'GET':
         form = SuiteForm(instance=obj)
         if request.session.get('status', None) == 'success':
@@ -66,7 +72,11 @@ def suite(request, pk):
     elif request.method == 'POST':
         creator = obj.creator
         form = SuiteForm(data=request.POST, instance=obj)
-        m2m_list = json.loads(request.POST.get('case', '[]'))
+        try:
+            m2m_list = json.loads(request.POST.get('case', 'null'))
+        except json.decoder.JSONDecodeError:
+            logger.warning('无法获取m2m值', exc_info=True)
+            m2m_list = None
         if form.is_valid():
             form_ = form.save(commit=False)
             form_.creator = creator
@@ -74,20 +84,21 @@ def suite(request, pk):
             form_.is_active = obj.is_active
             form_.save()
             # form.save_m2m()
-            m2m = SuiteVsCase.objects.filter(suite=obj).order_by('order')
-            original_m2m_list = list()
-            for dict_ in m2m.values('case'):
-                original_m2m_list.append(str(dict_['case']))
-            if original_m2m_list != m2m_list:
-                m2m.delete()
-                order = 0
-                for m2m_pk in m2m_list:
-                    if m2m_pk.strip() == '':
-                        continue
-                    order += 1
-                    m2m_object = Case.objects.get(pk=m2m_pk)
-                    SuiteVsCase.objects.create(suite=obj, case=m2m_object, order=order, creator=request.user,
-                                               modifier=request.user)
+            if m2m_list:
+                m2m = SuiteVsCase.objects.filter(suite=obj).order_by('order')
+                original_m2m_list = list()
+                for dict_ in m2m.values('case'):
+                    original_m2m_list.append(str(dict_['case']))
+                if original_m2m_list != m2m_list:
+                    m2m.delete()
+                    order = 0
+                    for m2m_pk in m2m_list:
+                        if m2m_pk.strip() == '':
+                            continue
+                        order += 1
+                        m2m_object = Case.objects.get(pk=m2m_pk)
+                        SuiteVsCase.objects.create(suite=obj, case=m2m_object, order=order, creator=request.user,
+                                                   modifier=request.user)
             request.session['status'] = 'success'
             redirect = request.POST.get('redirect')
             redirect_url = request.POST.get('redirect_url', '')
@@ -96,15 +107,16 @@ def suite(request, pk):
             else:
                 return HttpResponseRedirect(redirect_url)
         else:
-            # 暂存step列表
-            m2m = SuiteVsCase.objects.filter(suite=obj).order_by('order')
-            original_m2m_list = list()
-            for dict_ in m2m.values('case'):
-                original_m2m_list.append(str(dict_['case']))
-            temp_list = list()
-            temp_dict = dict()
-            if original_m2m_list != m2m_list:
-                temp_list_json = json.dumps(m2m_list)
+            if m2m_list:
+                # 暂存step列表
+                m2m = SuiteVsCase.objects.filter(suite=obj).order_by('order')
+                original_m2m_list = list()
+                for dict_ in m2m.values('case'):
+                    original_m2m_list.append(str(dict_['case']))
+                temp_list = list()
+                temp_dict = dict()
+                if original_m2m_list != m2m_list:
+                    temp_list_json = json.dumps(m2m_list)
         redirect_url = request.POST.get('redirect_url', '')
         is_success = False
         return render(request, 'main/suite/detail.html', locals())
@@ -121,7 +133,11 @@ def suite_add(request):
         return render(request, 'main/suite/detail.html', locals())
     elif request.method == 'POST':
         form = SuiteForm(data=request.POST)
-        m2m_list = json.loads(request.POST.get('case', '[]'))
+        try:
+            m2m_list = json.loads(request.POST.get('case', 'null'))
+        except json.decoder.JSONDecodeError:
+            logger.warning('无法获取m2m值', exc_info=True)
+            m2m_list = None
         if form.is_valid():
             form_ = form.save(commit=False)
             form_.creator = request.user
@@ -131,14 +147,15 @@ def suite_add(request):
             # form.save_m2m()
             obj = form_
             pk = obj.id
-            order = 0
-            for m2m_pk in m2m_list:
-                if m2m_pk.strip() == '':
-                    continue
-                order += 1
-                m2m_object = Case.objects.get(pk=m2m_pk)
-                SuiteVsCase.objects.create(suite=obj, case=m2m_object, order=order, creator=request.user,
-                                           modifier=request.user)
+            if m2m_list:
+                order = 0
+                for m2m_pk in m2m_list:
+                    if m2m_pk.strip() == '':
+                        continue
+                    order += 1
+                    m2m_object = Case.objects.get(pk=m2m_pk)
+                    SuiteVsCase.objects.create(suite=obj, case=m2m_object, order=order, creator=request.user,
+                                               modifier=request.user)
             request.session['status'] = 'success'
             redirect = request.POST.get('redirect')
             redirect_url = request.POST.get('redirect_url', '')
@@ -149,7 +166,8 @@ def suite_add(request):
             else:
                 return HttpResponseRedirect(redirect_url)
         else:
-            temp_list_json = json.dumps(m2m_list)
+            if m2m_list:
+                temp_list_json = json.dumps(m2m_list)
         redirect_url = request.POST.get('redirect_url', '')
         is_success = False
         return render(request, 'main/suite/detail.html', locals())
@@ -159,47 +177,38 @@ def suite_add(request):
 def suite_delete(request, pk):
     if request.method == 'POST':
         Suite.objects.filter(pk=pk).update(is_active=False, modifier=request.user, modified_date=timezone.now())
-        return HttpResponse('success')
+        return JsonResponse({'statue': 1, 'message': 'OK', 'data': pk})
     else:
-        return HttpResponseBadRequest('only accept "POST" method')
+        return JsonResponse({'statue': 2, 'message': 'Only accept "POST" method', 'data': pk})
 
 
 @login_required
 def suite_quick_update(request, pk):
     if request.method == 'POST':
-        response_ = {'new_value': ''}
         try:
             col_name = request.POST['col_name']
             new_value = request.POST['new_value']
-            response_['new_value'] = new_value
             obj = Suite.objects.get(pk=pk)
-            obj.modifier = request.user
-            obj.modified_date = timezone.now()
-            if col_name == 'name':
-                obj.name = new_value
+            if col_name in ('name', 'keyword'):
+                setattr(obj, col_name, new_value)
                 obj.clean_fields()
-                obj.save()
-            elif col_name == 'keyword':
-                obj.keyword = new_value
-                obj.clean_fields()
+                obj.modifier = request.user
+                obj.modified_date = timezone.now()
                 obj.save()
             else:
-                raise ValueError('invalid col_name')
+                raise ValueError('非法的字段名称')
         except Exception as e:
-            print(traceback.format_exc())
-            return HttpResponseBadRequest(str(e))
-        return JsonResponse(response_)
+            return JsonResponse({'statue': 2, 'message': str(e), 'data': None})
+        return JsonResponse({'statue': 1, 'message': 'OK', 'data': new_value})
     else:
-        return HttpResponseBadRequest('only accept "POST" method')
+        return JsonResponse({'statue': 2, 'message': 'Only accept "POST" method', 'data': None})
 
 
 # 获取选中的case
 @login_required
 def suite_cases(request, pk):
-    v = Case.objects.filter(suite=pk, is_active=True).order_by('suitevscase__order').values('pk', 'name', order=F('suitevscase__order'))
-    data_dict = dict()
-    data_dict['data'] = list(v)
-    return JsonResponse(data_dict)
+    objects = Case.objects.filter(suite=pk, is_active=True).order_by('suitevscase__order').values('pk', 'name', order=F('suitevscase__order'))
+    return JsonResponse({'statue': 1, 'message': 'OK', 'data': list(objects)})
 
 
 # 获取临时case
@@ -217,14 +226,16 @@ def case_list_temp(request):
         objects = list(objects)
         objects[0]['order'] = order
         list_temp.append(objects[0])
-    data_dict = dict()
-    data_dict['data'] = list_temp
-    return JsonResponse(data_dict)
+    return JsonResponse({'statue': 1, 'message': 'OK', 'data': list_temp})
 
 
 # 执行套件
 @login_required
 def suite_execute(request, pk):
-    from py_test.general.execute_suite import execute_suite
-    suite_result = execute_suite(request, pk, 'result1111')
-    return JsonResponse(model_to_dict(suite_result))
+    try:
+        suite_ = Suite.objects.get(pk=pk, is_active=True)
+        from py_test.general.execute_suite import execute_suite
+        suite_result = execute_suite(request, suite_, 'result1111')
+        return JsonResponse({'statue': 1, 'message': 'OK', 'data': model_to_dict(suite_result)})
+    except Suite.DoesNotExist:
+        return JsonResponse({'statue': 2, 'message': 'Suite does not exist', 'data': None})
